more exceptions handled

+++ main.py
@@ -226,14 +226,22 @@ class MainWindow(QMainWindow):
                         print(conflicts)
                         self.copy(source_path, target_path, conflicts)
                     else:
-                        new_location = self.copy_input(item.text())
-                        print(new_location)
-                        if new_location is not None:
-                            new_path = os.path.join(target_file_window.path, new_location.strip('[]'))
-                            if new_path != target_path:
-                                target_path = new_path
-                                conflicts = functions.compare_2_directories(source_path + "\\", target_path)
-                                self.copy(source_path, target_path, conflicts)
+                        while True:
+                            new_location = self.copy_input(item.text())
+                            print(new_location)
+                            if new_location is not None:
+                                contains_illegal_chars = any(char in new_location for char in ['/', '\\', ':', '*', '?', '"', '<', '>', '|'])
+                            else:
+                                break
+                            if new_location is not None and contains_illegal_chars is False and len(new_location) > 0 and new_location[-1] != ' ' and new_location[0] != ' ':
+                                new_path = os.path.join(target_file_window.path, new_location.strip('[]'))
+                                if new_path != target_path:
+                                    target_path = new_path
+                                    conflicts = functions.compare_2_directories(source_path + "\\", target_path)
+                                    self.copy(source_path, target_path, conflicts)
+                                    break
+                            else:
+                                QMessageBox.information(self, "Message", "Name can't contain / \\ : * ? \" < > | and cannot start or end with a space")
                 else:
                     print("isfile")
                     if source_path != target_path:
@@ -247,14 +255,23 @@ class MainWindow(QMainWindow):
                         else:
                             shutil.copy2(source_path, target_path)
                     else:
-                        new_location = self.copy_input(item.text())
-                        print(new_location)
-                        if new_location is not None:
-                            new_path = os.path.join(target_file_window.path, new_location.strip('[]'))
-                            if new_path != target_path:
-                                target_path = new_path
-                                shutil.copy2(source_path, target_path)
+                        while True:
+                            new_location = self.copy_input(item.text())
+                            print(new_location)
+                            if new_location is not None:
+                                contains_illegal_chars = any(char in new_location for char in ['/', '\\', ':', '*', '?', '"', '<', '>', '|'])
+                            else:
+                                break
+                            if new_location is not None and contains_illegal_chars is False and len(new_location) > 0 and new_location[-1] != ' ' and new_location[0] != ' ':
+                                new_path = os.path.join(target_file_window.path, new_location.strip('[]'))
+                                if new_path != target_path:
+                                    target_path = new_path
+                                    shutil.copy2(source_path, target_path)
+                                break
+                            else:
+                                QMessageBox.information(self, "Message", "Name can't contain / \\ : * ? \" < > | and cannot start or end with a space")
             target_file_window.display_directory_contents(target_file_window.path)  # Refresh the target file window
+            active_file_window.display_directory_contents(active_file_window.path)
         elif event.key() == Qt.Key_F6:  # Check if F6 is pressed
             print("moving")
             active_file_window = self.get_active_file_window()
@@ -276,15 +293,23 @@ class MainWindow(QMainWindow):
                         shutil.rmtree(source_path)
                         #os.rmdir(source_path)
                     else:
-                        new_location = self.move_input(item.text())
-                        print(new_location)
-                        if new_location is not None:
-                            new_path = os.path.join(target_file_window.path, new_location.strip('[]'))
-                            if new_path != target_path:
-                                target_path = new_path
-                                conflicts = functions.compare_2_directories(source_path + "\\", target_path)
-                                self.copy(source_path, target_path, conflicts)
-                                shutil.rmtree(source_path)
+                        while True:
+                            new_location = self.move_input(item.text())
+                            print(new_location)
+                            if new_location is not None:
+                                contains_illegal_chars = any(char in new_location for char in ['/', '\\', ':', '*', '?', '"', '<', '>', '|'])
+                            else:
+                                break
+                            if new_location is not None and contains_illegal_chars is False and len(new_location) > 0 and new_location[-1] != ' ' and new_location[0] != ' ':
+                                new_path = os.path.join(target_file_window.path, new_location.strip('[]'))
+                                if new_path != target_path:
+                                    target_path = new_path
+                                    conflicts = functions.compare_2_directories(source_path + "\\", target_path)
+                                    self.copy(source_path, target_path, conflicts)
+                                    shutil.rmtree(source_path)
+                                    break
+                            else:
+                                QMessageBox.information(self, "Message", "Name can't contain / \\ : * ? \" < > | and cannot start or end with a space")
                 else:
                     print("isfile")
                     if source_path != target_path:
@@ -301,14 +326,23 @@ class MainWindow(QMainWindow):
                             shutil.copy2(source_path, target_path)
                             os.remove(source_path)
                     else:
-                        new_location = self.move_input(item.text())
-                        print(new_location)
-                        if new_location is not None:
-                            new_path = os.path.join(target_file_window.path, new_location.strip('[]'))
-                            if new_path != target_path:
-                                target_path = new_path
-                                shutil.copy2(source_path, target_path)
-                                os.remove(source_path)
+                        while True:
+                            new_location = self.move_input(item.text())
+                            print(type(new_location))
+                            print(new_location is None)
+                            if new_location is not None:
+                                contains_illegal_chars = any(char in new_location for char in ['/', '\\', ':', '*', '?', '"', '<', '>', '|'])
+                            else:
+                                break
+                            if new_location is not None and contains_illegal_chars is False and len(new_location) > 0 and new_location[-1] != ' ' and new_location[0] != ' ':
+                                new_path = os.path.join(target_file_window.path, new_location.strip('[]'))
+                                if new_path != target_path:
+                                    target_path = new_path
+                                    shutil.copy2(source_path, target_path)
+                                    os.remove(source_path)
+                                    break
+                            else:
+                                QMessageBox.information(self, "Message", "Name can't contain / \\ : * ? \" < > | and cannot start or end with a space")
             target_file_window.display_directory_contents(target_file_window.path)  # Refresh the target file window
             active_file_window.display_directory_contents(active_file_window.path)
